# [stdlib] Constrain BinaryInteger.Words to be a Collection