--- a/stdlib/public/core/DoubleWidth.swift.gyb
+++ b/stdlib/public/core/DoubleWidth.swift.gyb
@@ -234,7 +234,8 @@ public struct DoubleWidth<Base : FixedWidthInteger> :
       if Base.bitWidth < UInt.bitWidth {
         _precondition(i == Index(.low(_low.startIndex)), "Invalid index")
         _sanityCheck(2 * Base.bitWidth <= UInt.bitWidth)
-        return _low.first! | (_high.first! &<< Base.bitWidth._lowWord) 
+        return _low[_low.startIndex] | 
+          (_high[_high.startIndex] &<< Base.bitWidth._lowWord) 
       }
       switch i._value {
       case let .low(li): return _low[li]