<!-- What's in this pull request? -->

`BinaryInteger.Words` could previously only be constrained to a `Sequence` of `UInts`; it is now possible to constrain it to `Collection`, so that we match SE-104.

<!-- If this pull request resolves any bugs in the Swift bug tracker, provide a link: -->

<!--
Before merging this pull request, you must run the Swift continuous integration tests.
For information about triggering CI builds via @swift-ci, see:
https://github.com/apple/swift/blob/master/docs/ContinuousIntegration.md#swift-ci

Thank you for your contribution to Swift!
-->