@moiseev 2017-07-25
Can it be a stronger requirement, like a `BidirectionalCollection` for example?