@lorentey 2017-07-25
@moiseev Not yet (or at least not as easily): it leads to seemingly unrelated errors in BidirectionalCollection.swift and elsewhere, and then a compiler crash. I'll see if I can make sense of it.

It is also not yet possible to have `Int.Words` and `UInt.Words` conform to `RandomAccessCollection` -- so we can't replace the custom implementation with `CollectionOfOne` yet.