@moiseev 2017-07-25
@lorentey I don't understand your second point about inability to use `CollectionOfOne`. It is still a `Collection` so regardless of the constraints on `Words` it should be usable... What am I missing?