@lorentey 2017-07-25
@moiseev Me neither! Something gets confused when `[U]Int.Words` is a random-access collection, but it's okay for `Int8.Words` and the rest of the integers. However, luckily we now get error messages rather than just a crash:

```
/Users/lorentey/Git/Swift/swift/stdlib/public/core/Indices.swift.gyb:127:17: error: 'Self.Indices.Indices.IndexDistance.Words.SubSequence.Element' cannot be equal to both 'Base.Element' and 'UInt'
  where Indices == DefaultBidirectionalIndices<Self>
                ^
/Users/lorentey/Git/Swift/swift/stdlib/public/core/Indices.swift.gyb:127:17: note: same-type constraint 'Self.Indices.Indices.IndexDistance.Words.Iterator.Element' == 'UInt' implied here
  where Indices == DefaultBidirectionalIndices<Self>
                ^
/Users/lorentey/Git/Swift/swift/stdlib/public/core/Indices.swift.gyb:127:17: error: 'Self.Indices.Indices.IndexDistance.Words.SubSequence.Element' cannot be equal to both 'Base.Element' and 'UInt'
  where Indices == DefaultBidirectionalIndices<Self>
                ^
/Users/lorentey/Git/Swift/swift/stdlib/public/core/Indices.swift.gyb:127:17: note: same-type constraint 'Self.Indices.Indices.IndexDistance.Words.Iterator.Element' == 'UInt' implied here
  where Indices == DefaultBidirectionalIndices<Self>
                ^
/Users/lorentey/Git/Swift/swift/stdlib/public/core/Indices.swift.gyb:127:17: error: 'Self.Indices.Indices.IndexDistance.Words.SubSequence.Iterator.Element' cannot be equal to both 'Elements.Element' and 'UInt'
  where Indices == DefaultBidirectionalIndices<Self>
                ^
```

(It goes on for a couple dozen more errors like this.) Does this ring a bell for you?