@moiseev 2017-07-25
> /Users/lorentey/Git/Swift/swift/stdlib/public/core/Indices.swift.gyb:127:17: error: 'Self.Indices.Indices.IndexDistance.Words.SubSequence.Element' cannot be equal to both 'Base.Element' and 'UInt'

This begs for a another equality constraint like `SubSequence.Base.Element == UInt`. But hard to tell... And no, I haven't seen anything like that before.