@moiseev 2017-07-25
@swift-ci Please smoke benchmark