@moiseev 2017-07-25
Since messing with constraints like this sometimes has an unwelcome changes in compilation time, I'd rather not merge this change without measuring the stdlib compile time with and without this patch applied.