@lorentey 2017-07-25
Understood! 👍 Shall I time it on my end?

FWIW, my timings for incremental builds using `./swift/utils/build-script -R` aren't great:

```
before: 264.68 real      1099.82 user        48.22 sys
after:  301.12 real      1223.43 user        53.47 sys
```