@moiseev 2017-07-27
@swift-ci Please smoke test