@lorentey 2017-07-27
I fixed the build time regression by relaxing the BinaryInteger.Words constraint to `_Indexable`.

To ensure `words.count` remains available in generic contexts, I moved `Collection`'s `count` property up into `_Indexable` (along with `isEmpty`).

Updated build times: (incremental `build-script -R` on latest master)

```
before: 269.12 real      1091.76 user        47.95 sys
after:  274.38 real      1136.65 user        49.71 sys
```