@moiseev 2017-07-28
@swift-ci Please test