@moiseev 2017-07-31
FWIW, here are my local measurements of build time with and without the change applied:

```
before: 317.25s user 25.48s system 303% cpu 1:53.08 total
after: 340.38s user 26.32s system 301% cpu 2:01.79 total

before: 309.81s user 25.82s system 290% cpu 1:55.43 total
after: 343.30s user 26.98s system 287% cpu 2:08.89 total
```

Regression of about ~10%.