@lorentey 2017-08-01
@moiseev Weird; I get consistently better ratios. (On an i7 rMBP with -j8, with the PR rebased onto current master.)

With `Collection` conformance:

 Name | real | user | sys
 --- | --- | --- | ---
 before  |      295.93  |    1234.35  |      64.31
 after c809110  |     332.85   |   1380.12   |     68.63
 Change | 112% | 112% | 107%

With `_Indexable`:

 Name | real | user | sys
 --- | --- | --- | ---
 before  |      295.93  |    1234.35  |      64.31
 after 11e30ed  |     304.44  |    1282.90  |      63.69
 Change | 103% | 104% | 99%

In any case, instance-level `bitWidth` is still in the `BinaryInteger` protocol, so people can continue to use that in generic contexts; it's not super important to make `words.count` available, too. I'm a-okay with closing this PR without a merge! (I'll keep monitoring compiler performance and resubmit if/when we can do this without a performance penalty.)