@moiseev 2017-08-04
@lorentey what are you timing exactly? I just cannot explain why I see the difference and you don't =)

My procedure is essentially this:

```sh
utils/build-script -r --skip-build-benchmarks
find ./stdlib/public/core/* -name '*.swift*' -print0 | xargs -0 touch
time ninja -C ../build/Ninja-RelWithDebInfoAssert+stdlib-DebugAssert/swift-macosx-x86_64 swift-stdlib-macosx
```

What's yours?

I tried it again just now on my desktop with these results:

```

before 147.09s user 8.52s system 240% cpu 1:04.63 total
after 162.12s user 8.68s system 245% cpu 1:09.68 total
```