@lorentey 2017-08-04
@moiseev Ah, that explains it! I just used `time build-script -R`, so my timings included the compilation time for benchmarks and overlays, too. 🐢 (Ninja compiled about 510 or so files, AFAICR.)